requirements.txt
anthropic>=0.40.0
openai>=1.50.0
datasets>=2.18.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
